Make the new version check script more robust.
* github/scripts/check-new-version.py (test_version): Drop the "esr"
suffix from the version strings before feeding them to version.parse,
since it only supports versions that follow the PEP 440 requirements.
Also, only try to get the build number from current_version if it is
actually present (it may not be, if we changed current_version to one
of the new candidate versions as we try to find the latest one).