Bump LLVM to 15.0.6 on amd64
Building Firefox beta 110 with LLVM 14.0.0 has been failing on amd64
with the following error[1] (lines wrapped and indented):

    error: linking with
    `/build/firefox/parts/firefox/build/build/cargo-linker` failed:
    exit status: 1

    note: /snap/gnome-3-38-2004-sdk/current/usr/bin/ld.gold:
    fatal error: LLVM gold plugin has failed to create LTO module:
    Opaque pointers are only supported in -opaque-pointers mode
    (Producer: 'LLVM15.0.0-rust-1.65.0-stable' Reader: 'LLVM 14.0.0')

Bumping LLVM to 15 seems to fix it.  However, we currently do it only
on amd64, as building with LLVM 15 seems to currently fail on arm64,
at least when using the official LLVM 15.0.6 binaries provided by the
LLVM project.

[1] Sample build for reference:
https://launchpad.net/~mozilla-snaps/firefox/+snap/firefox-snap-beta/+build/2000291