snap, hooks: make host-hunspell a system-files interface
The host-hunspell plug was a mount-control interface, which means that the snap
could request snapd to set up a mount from/to a specific location. The problem
is that the mount is visible in the host mount namespace, i.e. the mounts
created a system wide. The mount-control interface was really intended to be
used in Ubuntu Core oriented scenarios, when for instance a gadget mounts some
external storage at a system wide location.

The particular care of the Firefox snap only really needs access to the hunspell
dictionaries on the host. In this case a system-files interface allowing access
to /usr/share/hunspell inside the host filesystem (exposed to the snap mount ns
at /var/lib/snapd/hostfs) will be sufficient.

This will fix issues like LP#2059195

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>