Bug 1899701 - Add home-read-all plug
Add a 'home-read-all' plug that sets the 'read' attribute for the
'home' interface to the special value 'all', to allow the Firefox snap
to read files in the user's home directory that are owned by another
user but that the current user should still be able to read via group
membership.

Note: 'read: all' technically allows reading the home directory of any
user, but per snapd folks that's as granular as we can get because
AppArmor does not have per-user profiles.

https://bugzilla.mozilla.org/1899701
https://snapcraft.io/docs/home-interface